Exit point problem
You are given n*m numbers (1's and 0's), representing elements of 2d array a. Consider this array a maze and a player enters from top-left corner in east direction.The player moves in the same direction as long as he meets '0'. On seeing a 1, he takes a 90 deg right turn.