Paint House - Many colors
You are given a number n and a number k separated by a space, representing the number of houses and number of colors.In the next n rows, you are given k space separated numbers representing the cost of painting nth house with one of the k colors. You are required to calculate and print the minimum cost of painting all houses without painting any consecutive house with same color.